Second attempt to fix shadow leaks (which is successfull!)
I thought of this in the morning and it actually worked. Basically, discard the bake point if most rays land on backfaces and use the neighbors to fill it. This is done in the dilating method, but I'll likely merge it with the paint function in the future.

I added gamma correction as well. Since it balances the brightness better, enclosed areas aren't pitch black now.